requirements/midcare_app.txt
django
djangorestframework
mysqlclient
django-storages
boto3
django-environ